upload cell line counts
