+++ process-dataset/index.js
@@ -4,13 +4,14 @@ const uploadDatasetAndManifest = require("./upload-dataset-and-manifest");
 const uploadFeatureDefs = require("./upload-feature-defs");
 const uploadCellLines = require("./upload-cell-lines");
 const formatAndWritePerCellJsons = require("./write-per-cell-jsons");
+const uploadCellCountsPerCellLine = require("./upload-cell-counts");
 const uploadFileInfo = require("./upload-file-info");
 const uploadFileToS3 = require("./upload-to-aws");
 
 const FirebaseHandler = require('./firebase/firebase-handler');
 
 const args = process.argv.slice(2);
-console.log('myArgs: ', args);
+console.log('Received: ', args);
 
 const datasetReadFolder = args[0];
 
@@ -40,18 +41,20 @@ const processDataset = async () => {
     const firebaseHandler = new FirebaseHandler(id);
 
     // 1. upload dataset description and manifest
-    // await uploadDatasetAndManifest(firebaseHandler, datasetJson, datasetReadFolder)
+    await uploadDatasetAndManifest(firebaseHandler, datasetJson, datasetReadFolder)
     // 2. check dataset feature defs for new features, upload them if needed
-    // await uploadFeatureDefs(firebaseHandler, datasetReadFolder)
+    await uploadFeatureDefs(firebaseHandler, datasetReadFolder)
     // 3. upload cell lines TODO: add check if cell line is already there
-    // await uploadCellLines(firebaseHandler, datasetReadFolder)
+    await uploadCellLines(firebaseHandler, datasetReadFolder)
     // 4. format file info, write to json locally
     await formatAndWritePerCellJsons(datasetReadFolder, "./data");
-    // 5. upload file info per cell
+    // 5. upload cell line subtotals
+    await uploadCellCountsPerCellLine("./data", firebaseHandler)
+    // 6. upload file info per cell
     const fileInfoLocation = await uploadFileInfo(firebaseHandler, "./data");
-    // 6. upload json to aw3
+    // 7. upload json to aw3
     const awsLocation = await uploadFileToS3(id, "./data");
-    // 7. update dataset manifest with location for data
+    // 8. update dataset manifest with location for data
     const updateToManifest = {
         ...fileInfoLocation, 
         ...awsLocation
